extract unused bindings to their own file + extracted preloader to its own file, so to shring fileupload bindings.
