+++ src/js/ko-bindings.js
@@ -9,6 +9,7 @@ require("./bindings/blocks.js");
 require("./bindings/csstext.js");
 require("./bindings/bind-iframe.js");
 require("./bindings/extdroppable.js");
+require("./bindings/preloader.js");
 require("./bindings/fileupload.js");
 require("./bindings/virtuals.js");
 require("./bindings/wysiwygs.js");
